more S3Provider cleanup
