+++ src/concurrency/sync/CloudRotateKeyTask.cpp
@@ -1,5 +1,5 @@
 #include "concurrency/sync/CloudRotateKeyTask.hpp"
-#include "storage/CloudStorageEngine.hpp"
+#include "storage/cloud/CloudStorageEngine.hpp"
 #include "types/File.hpp"
 #include "types/RSync.hpp"
 #include "database/Queries/FileQueries.hpp"
