+++ src/concurrency/sync/MirrorSyncTask.cpp
@@ -1,8 +1,5 @@
 #include "concurrency/sync/MirrorSyncTask.hpp"
-#include "concurrency/ThreadPool.hpp"
-#include "database/Queries/FileQueries.hpp"
-#include "storage/StorageManager.hpp"
-#include "storage/CloudStorageEngine.hpp"
+#include "storage/cloud/CloudStorageEngine.hpp"
 #include "types/File.hpp"
 #include "types/Directory.hpp"
 #include "types/RSync.hpp"
