+++ src/concurrency/sync/UploadTask.cpp
@@ -1,5 +1,5 @@
 #include "concurrency/sync/UploadTask.hpp"
-#include "storage/CloudStorageEngine.hpp"
+#include "storage/cloud/CloudStorageEngine.hpp"
 #include "types/File.hpp"
 #include "services/LogRegistry.hpp"
