+++ src/storage/cloud/s3/file.cpp
@@ -1,0 +1,136 @@
+#include "storage/cloud/s3/S3Controller.hpp"
+#include "services/LogRegistry.hpp"
+
+#include <fstream>
+
+using namespace vh::cloud;
+using namespace vh::types;
+using namespace vh::util;
+using namespace vh::logging;
+
+void S3Controller::uploadLargeObject(const std::filesystem::path& key,
+const std::filesystem::path& filePath,
+const uintmax_t partSize) const {
+    std::ifstream file(filePath, std::ios::binary);
+    if (!file) throw std::runtime_error("Failed to open file for large upload: " + filePath.string());
+
+    const std::string uploadId = initiateMultipartUpload(key);
+    if (uploadId.empty()) throw std::runtime_error("Failed to initiate multipart upload for: " + key.string());
+
+    std::vector<std::string> etags;
+    int partNo = 1;
+    bool success = true;
+
+    while (file) {
+        std::string part(partSize, '\0');
+        file.read(&part[0], static_cast<std::streamsize>(partSize));
+        std::streamsize bytesRead = file.gcount();
+        if (bytesRead <= 0) break;
+
+        part.resize(static_cast<size_t>(bytesRead));
+
+        std::string etag;
+        try { uploadPart(key, uploadId, partNo++, part, etag); }
+        catch (const std::exception& e) {
+            LogRegistry::cloud()->error("[S3Provider] uploadLargeObject failed to upload part {}: {}", partNo - 1, e.what());
+            success = false;
+            break;
+        }
+        etags.push_back(std::move(etag));
+    }
+
+    if (!success || etags.empty()) {
+        try {
+            abortMultipartUpload(key, uploadId);
+            return;
+        }
+        catch (const std::exception& e) {
+            LogRegistry::cloud()->error(
+                "[S3Provider] uploadLargeObject failed to abort multipart upload for {}: uploadId={}",
+                key.string(), uploadId);
+            LogRegistry::cloud()->error(e.what());
+            throw std::runtime_error("Failed to abort multipart upload after part upload failure");
+        }
+    }
+
+    return completeMultipartUpload(key, uploadId, etags);
+}
+
+void S3Controller::uploadObject(const std::filesystem::path& key, const std::filesystem::path& filePath) const {
+    std::ifstream fin(filePath, std::ios::binary);
+    if (!fin) throw std::runtime_error("Failed to open file for upload: " + filePath.string());
+
+    fin.seekg(0, std::ios::end);
+    const curl_off_t sz = fin.tellg();
+    fin.seekg(0);
+
+    const std::string fileContents = slurp(fin);
+    const std::string payloadHash = sha256Hex(fileContents);
+
+    fin.clear();
+    fin.seekg(0);
+
+    CurlEasy tmpHandle;
+    const auto [canonical, url] = constructPaths(static_cast<CURL*>(tmpHandle), key);
+
+    SList hdrs = makeSigHeaders("PUT", canonical, payloadHash);
+    hdrs.add("Content-Type: application/octet-stream");
+
+    HttpResponse resp = performCurl([&](CURL* h) {
+        curl_easy_setopt(h, CURLOPT_URL, url.c_str());
+        curl_easy_setopt(h, CURLOPT_UPLOAD, 1L);
+        curl_easy_setopt(h, CURLOPT_HTTPHEADER, hdrs.get());
+        curl_easy_setopt(h, CURLOPT_READDATA, &fin);
+        curl_easy_setopt(h, CURLOPT_INFILESIZE_LARGE, sz);
+        curl_easy_setopt(h, CURLOPT_READFUNCTION,
+            +[](char* buf, const size_t sz, const size_t nm, void* ud) -> size_t {
+                auto* fp = static_cast<std::ifstream*>(ud);
+                fp->read(buf, static_cast<std::streamsize>(sz * nm));
+                return static_cast<size_t>(fp->gcount());
+            });
+    });
+
+    if (!resp.ok()) throw std::runtime_error(
+        fmt::format("Failed to upload file to S3 (HTTP {}): {}", resp.http, resp.body));
+}
+
+void S3Controller::downloadObject(const std::filesystem::path& key,
+                                const std::filesystem::path& outputPath) const {
+    CURL* curl = curl_easy_init();
+    if (!curl) throw std::runtime_error("Failed to init curl for S3 download");
+
+    std::ofstream file(outputPath, std::ios::binary);
+    if (!file) {
+        curl_easy_cleanup(curl);
+        throw std::runtime_error("Failed to open output file for S3 download: " + outputPath.string());
+    }
+
+    const auto [canonicalPath, url] = constructPaths(curl, key);
+
+    const std::string payloadHash = "UNSIGNED-PAYLOAD";
+    auto hdrMap = buildHeaderMap(payloadHash);
+    const std::string authHeader =  buildAuthorizationHeader(apiKey_, "GET", canonicalPath, hdrMap, payloadHash);
+
+    HeaderList headers;
+    headers.add("Authorization: " + authHeader);
+    for (const auto& [k, v] : hdrMap) headers.add(fmt::format("{}: {}", k, v));
+
+    auto writeFn = +[](const char* ptr, const size_t size, const size_t nmemb, void* userdata) -> size_t {
+        auto* fout = static_cast<std::ofstream*>(userdata);
+        fout->write(ptr, size * nmemb);
+        return size * nmemb;
+    };
+
+    curl_easy_setopt(curl, CURLOPT_URL, url.c_str());
+    curl_easy_setopt(curl, CURLOPT_HTTPHEADER, headers.list);
+    curl_easy_setopt(curl, CURLOPT_FOLLOWLOCATION, 1L);
+    curl_easy_setopt(curl, CURLOPT_WRITEFUNCTION, writeFn);
+    curl_easy_setopt(curl, CURLOPT_WRITEDATA, &file);
+
+    const CURLcode res = curl_easy_perform(curl);
+    curl_easy_cleanup(curl);
+    file.close();
+
+    if (res != CURLE_OK) throw std::runtime_error(
+        fmt::format("Failed to download file from S3: CURL error {}", res));
+}
