+++ src/storage/cloud/s3/multipart.cpp
@@ -1,0 +1,165 @@
+#include "storage/cloud/s3/S3Controller.hpp"
+#include "types/APIKey.hpp"
+#include "util/u8.hpp"
+#include "services/LogRegistry.hpp"
+
+#include <fstream>
+#include <regex>
+
+using namespace vh::cloud;
+using namespace vh::types;
+using namespace vh::util;
+using namespace vh::logging;
+
+std::string S3Controller::initiateMultipartUpload(const std::filesystem::path& key) const {
+    CURL* curl = curl_easy_init();
+    if (!curl) return "";
+
+    const auto keyStr = key.u8string();
+    const auto [canonicalPath, url] = constructPaths(curl, keyStr, "?uploads");
+    const std::string payloadHash = "UNSIGNED-PAYLOAD";
+
+    auto hdrMap = buildHeaderMap(payloadHash);
+    const std::string authHeader = buildAuthorizationHeader(apiKey_, "POST", canonicalPath, hdrMap, payloadHash);
+
+    HeaderList headers;
+    headers.add("Authorization: " + authHeader);
+    for (const auto& [k, v] : hdrMap) headers.add(k + ": " + v);
+
+    std::string response;
+    curl_easy_setopt(curl, CURLOPT_URL, url.c_str());
+    curl_easy_setopt(curl, CURLOPT_POST, 1L);
+    curl_easy_setopt(curl, CURLOPT_HTTPHEADER, headers.list);
+    curl_easy_setopt(curl, CURLOPT_POSTFIELDS, "");
+    curl_easy_setopt(curl, CURLOPT_POSTFIELDSIZE_LARGE, static_cast<curl_off_t>(0));
+    curl_easy_setopt(curl, CURLOPT_WRITEFUNCTION, writeToString);
+    curl_easy_setopt(curl, CURLOPT_WRITEDATA, &response);
+
+    CURLcode res = curl_easy_perform(curl);
+    long httpCode = 0;
+    curl_easy_getinfo(curl, CURLINFO_RESPONSE_CODE, &httpCode);
+    curl_easy_cleanup(curl);
+
+    if (res != CURLE_OK || httpCode != 200) {
+        LogRegistry::cloud()->error("[S3Provider] initiateMultipartUpload failed: CURL={} HTTP={} Response:\n{}",
+                                    res, httpCode, response);
+        return "";
+    }
+
+    std::smatch m;
+    std::regex re(R"(<UploadId>([^<]+)</UploadId>)");
+    if (std::regex_search(response, m, re) && m.size() > 1)
+        return m[1].str();
+
+    LogRegistry::cloud()->error("[S3Provider] initiateMultipartUpload failed to parse UploadId from response: {}", response);
+    return "";
+}
+
+void S3Controller::uploadPart(const std::filesystem::path& key, const std::string& uploadId,
+                             const int partNumber, const std::string& partData, std::string& etagOut) const {
+    CURL* curl = curl_easy_init();
+    if (!curl) throw std::runtime_error("Failed to init curl for S3 multipart upload");
+
+    const auto keyStr = key.u8string();
+    const std::string query = "?partNumber=" + std::to_string(partNumber) + "&uploadId=" + uploadId;
+    const auto [canonicalPath, url] = constructPaths(curl, keyStr, query);
+
+    const std::string payloadHash = sha256Hex(partData);
+    const auto hdrMap = buildHeaderMap(payloadHash);
+    const std::string authHeader = buildAuthorizationHeader(apiKey_, "PUT", canonicalPath, hdrMap, payloadHash);
+
+    HeaderList headers;
+    headers.add("Content-Type: application/octet-stream");
+    headers.add("Authorization: " + authHeader);
+    for (const auto& [k, v] : hdrMap) headers.add(k + ": " + v);
+
+    std::string respHdr;
+    curl_easy_setopt(curl, CURLOPT_URL, url.c_str());
+    curl_easy_setopt(curl, CURLOPT_CUSTOMREQUEST, "PUT");
+    curl_easy_setopt(curl, CURLOPT_HTTPHEADER, headers.list);
+    curl_easy_setopt(curl, CURLOPT_POSTFIELDS, partData.data());
+    curl_easy_setopt(curl, CURLOPT_POSTFIELDSIZE_LARGE, static_cast<curl_off_t>(partData.size()));
+    curl_easy_setopt(curl, CURLOPT_HEADERFUNCTION, writeToString);
+    curl_easy_setopt(curl, CURLOPT_HEADERDATA, &respHdr);
+
+    const CURLcode res = curl_easy_perform(curl);
+    long httpCode = 0;
+    curl_easy_getinfo(curl, CURLINFO_RESPONSE_CODE, &httpCode);
+    curl_easy_cleanup(curl);
+
+    if (res != CURLE_OK || httpCode != 200)
+        throw std::runtime_error(fmt::format("Failed to upload part {}: CURL={} HTTP={}", partNumber, res, httpCode));
+
+    if (!extractETag(respHdr, etagOut))
+        throw std::runtime_error(fmt::format("Failed to extract ETag for uploaded part {}", partNumber));
+}
+
+void S3Controller::completeMultipartUpload(const std::filesystem::path& key, const std::string& uploadId,
+                                         const std::vector<std::string>& etags) const {
+    if (etags.empty()) throw std::runtime_error("No ETags provided to completeMultipartUpload");
+
+    CURL* curl = curl_easy_init();
+    if (!curl) throw std::runtime_error("Failed to init curl for S3 complete multipart upload");
+
+    std::u8string keyStr = key.u8string();
+    const std::string query = "?uploadId=" + uploadId;
+    const auto [canonicalPath, url] = constructPaths(curl, keyStr, query);
+
+    const auto body = composeMultiPartUploadXMLBody(etags);
+    const std::string payloadHash = sha256Hex(body);
+    auto hdrMap = buildHeaderMap(payloadHash);
+    const std::string authHeader = buildAuthorizationHeader(apiKey_, "POST", canonicalPath, hdrMap, payloadHash);
+
+    HeaderList headers;
+    headers.add("Content-Type: application/xml");
+    headers.add("Authorization: " + authHeader);
+    for (const auto& [k, v] : hdrMap) headers.add(k + ": " + v);
+
+    std::string response;
+    curl_easy_setopt(curl, CURLOPT_URL, url.c_str());
+    curl_easy_setopt(curl, CURLOPT_CUSTOMREQUEST, "POST");
+    curl_easy_setopt(curl, CURLOPT_HTTPHEADER, headers.list);
+    curl_easy_setopt(curl, CURLOPT_POSTFIELDS, body.data());
+    curl_easy_setopt(curl, CURLOPT_POSTFIELDSIZE_LARGE, static_cast<curl_off_t>(body.size()));
+    curl_easy_setopt(curl, CURLOPT_WRITEFUNCTION, writeToString);
+    curl_easy_setopt(curl, CURLOPT_WRITEDATA, &response);
+
+    CURLcode res = curl_easy_perform(curl);
+    long httpCode = 0;
+    curl_easy_getinfo(curl, CURLINFO_RESPONSE_CODE, &httpCode);
+    curl_easy_cleanup(curl);
+
+    if (res != CURLE_OK || httpCode != 200) {
+        LogRegistry::cloud()->error("[S3Provider] completeMultipartUpload failed: CURL={} HTTP={} Response:\n{}",
+                                    res, httpCode, response);
+        throw std::runtime_error(
+            fmt::format("Failed to complete multipart upload to S3 (HTTP {}): {}", httpCode, response));
+    }
+}
+
+void S3Controller::abortMultipartUpload(const std::filesystem::path& key, const std::string& uploadId) const {
+    CURL* curl = curl_easy_init();
+    if (!curl) throw std::runtime_error("Failed to init curl for S3 abort multipart upload");
+
+    const auto keyStr = key.u8string();
+    const std::string query = "?uploadId=" + uploadId;
+    const auto [canonicalPath, url] = constructPaths(curl, keyStr, query);
+
+    const std::string payloadHash = sha256Hex("");
+    const auto hdrMap = buildHeaderMap(payloadHash);
+    const std::string authHeader = buildAuthorizationHeader(apiKey_, "DELETE", canonicalPath, hdrMap, payloadHash);
+
+    HeaderList headers;
+    headers.add("Authorization: " + authHeader);
+    for (const auto& [k, v] : hdrMap) headers.add(k + ": " + v);
+
+    curl_easy_setopt(curl, CURLOPT_URL, url.c_str());
+    curl_easy_setopt(curl, CURLOPT_CUSTOMREQUEST, "DELETE");
+    curl_easy_setopt(curl, CURLOPT_HTTPHEADER, headers.list);
+
+    const CURLcode res = curl_easy_perform(curl);
+    curl_easy_cleanup(curl);
+
+    if (res != CURLE_OK)
+        throw std::runtime_error("Failed to abort multipart upload to S3: CURL error " + std::to_string(res));
+}
