+++ tests/test_S3Provider.cpp
@@ -1,5 +1,5 @@
 #include "types/APIKey.hpp"
-#include "storage/cloud/S3Controller.hpp"
+#include "storage/cloud/s3/S3Controller.hpp"
 #include "util/imageUtil.hpp"
 #include "types/FSEntry.hpp"
 
@@ -11,11 +11,16 @@
 
 namespace fs = std::filesystem;
 
+using namespace vh::cloud;
+using namespace vh::types;
+using namespace vh::util;
+using namespace vh::types::api;
+
 class S3ProviderIntegrationTest : public ::testing::Test {
   protected:
-    std::shared_ptr<vh::types::api::APIKey> apiKey_;
+    std::shared_ptr<APIKey> apiKey_;
     std::string bucket_;
-    std::shared_ptr<vh::cloud::S3Controller> s3Provider_;
+    std::shared_ptr<S3Controller> s3Provider_;
     std::filesystem::path test_dir;
     bool skipTests = false;
 
@@ -32,9 +37,9 @@ class S3ProviderIntegrationTest : public ::testing::Test {
             return;
         }
 
-        apiKey_ = std::make_shared<vh::types::api::APIKey>(1,
+        apiKey_ = std::make_shared<APIKey>(1,
                                                              "Test S3 Key",
-                                                             vh::types::api::S3Provider::CloudflareR2,
+                                                             S3Provider::CloudflareR2,
                                                              std::getenv("VAULTHALLA_TEST_R2_ACCESS_KEY"),
                                                              std::getenv("VAULTHALLA_TEST_R2_SECRET_ACCESS_KEY"),
                                                              std::getenv("VAULTHALLA_TEST_R2_REGION"),
@@ -42,7 +47,7 @@ class S3ProviderIntegrationTest : public ::testing::Test {
 
         bucket_ = std::getenv("VAULTHALLA_TEST_R2_BUCKET");
 
-        s3Provider_ = std::make_shared<vh::cloud::S3Controller>(apiKey_, bucket_);
+        s3Provider_ = std::make_shared<S3Controller>(apiKey_, bucket_);
 
         FPDF_LIBRARY_CONFIG config;
         config.version = 3;
@@ -210,7 +215,7 @@ TEST_F(S3ProviderIntegrationTest, test_S3ListObjectsAndDownloadToBuffer) {
 
     const auto xml = s3Provider_->listObjects();
 
-    const auto entries = vh::types::fromS3XML(xml);
+    const auto entries = fromS3XML(xml);
     EXPECT_FALSE(entries.empty()) << "fromS3XML should return at least one entry";
     auto match = std::find_if(entries.begin(), entries.end(), [&](const auto& entry) {
         return !entry->isDirectory() && entry->path.filename() == key;
@@ -240,7 +245,7 @@ TEST_F(S3ProviderIntegrationTest, test_ResizeAndCompressImageBuffer) {
     std::vector<uint8_t> buffer;
     ASSERT_NO_THROW(s3Provider_->downloadToBuffer(key, buffer));
 
-    auto jpeg = vh::util::resize_and_compress_image_buffer(
+    auto jpeg = resize_and_compress_image_buffer(
         buffer.data(),
         buffer.size(),
         std::nullopt,
@@ -262,7 +267,7 @@ TEST_F(S3ProviderIntegrationTest, test_ResizeAndCompressPdfBuffer) {
     std::vector<uint8_t> buffer;
     ASSERT_NO_THROW(s3Provider_->downloadToBuffer(key, buffer));
 
-    auto jpeg = vh::util::resize_and_compress_pdf_buffer(
+    auto jpeg = resize_and_compress_pdf_buffer(
         buffer.data(),
         buffer.size(),
         std::nullopt,
